Move binaries script from provisionning to bin
I realize we have more and more scripts located in the provisionning directory.
But they are not all related to provisionning.
Lots of them are simple wrapper to avoid writing `docker-compose exec -T php <command>`.

The provisionning directory should only contain what is required to define our docker containers.
The bin directory is meant to store those kind of scripts.