[Symfony DI] Make command handlers services lazy (#13)
Command handlers are injected in the command bus.
It means that when you call the command_bus service, you will load every command handler in your application.
This is not a wanted behaviour.
We only use the command_bus to handle one command handler. Requiring every handler to be instantiated will make
our application slower than it can be.

The solution is to define every command handler services as lazy and (sadely) remove the final keyword from
command handler classes.